Seperate income from donations (PurdueIEEE/boilerbooks#159)
